fix(bt/bluedroid): fixed issues from AI review in GAP, SPP, HID, L2CAP and HCI (v5.2)

See merge request espressif/esp-idf!51082

+++ components/bt/host/bluedroid/stack/include/stack/hidd_api.h
@@ -111,7 +111,7 @@ extern tHID_STATUS HID_DevDeregister(void);
  * Returns          tHID_STATUS
  *
  ******************************************************************************/
-extern tHID_STATUS HID_DevSetSecurityLevel(uint8_t sec_lvl);
+extern tHID_STATUS HID_DevSetSecurityLevel(uint16_t sec_lvl);
 
 /*******************************************************************************
  *
